refactor: centralize test fixtures
