+++ Backend/tests/test_api.py
@@ -1,43 +1,4 @@
-import os
-import sys
-from typing import Iterator
-
-import pytest
 from fastapi.testclient import TestClient
-from sqlalchemy.pool import StaticPool
-from sqlmodel import Session, SQLModel, create_engine
-
-sys.path.append(os.path.dirname(os.path.dirname(os.path.dirname(__file__))))
-
-from Backend import models  # ensure models are imported for SQLModel metadata
-from Backend.backend import app
-from Backend.db import get_db
-
-
-@pytest.fixture(name="client")
-def client_fixture() -> Iterator[TestClient]:
-    database_url = os.environ.get("DATABASE_URL", "sqlite://")
-
-    engine_kwargs = {}
-    if database_url.startswith("sqlite"):
-        engine_kwargs = {
-            "connect_args": {"check_same_thread": False},
-            "poolclass": StaticPool,
-        }
-
-    engine = create_engine(database_url, **engine_kwargs)
-    SQLModel.metadata.create_all(engine)
-
-    def override_get_db() -> Iterator[Session]:
-        with Session(engine) as session:
-            yield session
-
-    app.dependency_overrides[get_db] = override_get_db
-    with TestClient(app) as client:
-        yield client
-
-    SQLModel.metadata.drop_all(engine)
-    app.dependency_overrides.clear()
 
 
 def test_ingredient_crud(client: TestClient) -> None:
